Moved PaletteCellIconEngine from palettetree.h to separate file
