+++ src/palette/internal/palette/palettetree.cpp
@@ -1060,190 +1060,4 @@ void PaletteTree::retranslate()
         p->retranslate();
     }
 }
-
-//---------------------------------------------------------
-//   paintIconElement
-/// Paint an icon element so that it fills a QRect, preserving
-/// aspect ratio, and leaving a small margin around the edges.
-//---------------------------------------------------------
-
-static void paintIconElement(mu::draw::Painter& painter, const RectF& rect, Element* e)
-{
-    Q_ASSERT(e && e->isActionIcon());
-    painter.save();
-
-    constexpr qreal margin = 4.0;
-    qreal extent = qMin(rect.height(), rect.width()) - margin;
-
-    ActionIcon* action = toActionIcon(e);
-    action->setExtent(extent);
-
-    extent /= 2.0;
-    PointF iconCenter(extent, extent);
-
-    painter.translate(rect.center() - iconCenter);
-    action->draw(&painter);
-    painter.restore();
-}
-
-//---------------------------------------------------------
-//   paintScoreElement
-/// Paint a non-icon element centered at the origin of the
-/// painter's coordinate system. If alignToStaff is true then
-/// the element is only centered horizontally; i.e. vertical
-/// alignment is unchanged from the default so that item will
-/// appear at the correct height on the staff.
-//---------------------------------------------------------
-
-void PaletteCellIconEngine::paintScoreElement(mu::draw::Painter& painter, Element* element, qreal spatium, bool alignToStaff) const
-{
-    Q_ASSERT(element && !element->isActionIcon());
-    painter.save(); // so we can restore painter after we are done using it
-
-    const qreal sizeRatio = spatium / gscore->spatium();
-    painter.scale(sizeRatio, sizeRatio);   // scale coordinates so element is drawn at correct size
-
-    element->layout(); // calculate bbox
-    PointF origin = element->bbox().center();
-
-    if (alignToStaff) {
-        // y = 0 is position of the element's parent.
-        // If the parent is the staff (or a segment on the staff) then
-        // y = 0 corresponds to the position of the top staff line.
-        origin.setY(0.0);
-    }
-
-    painter.translate(-1.0 * origin); // shift coordinates so element is drawn at correct position
-
-    element->scanElements(&painter, Palette::paintPaletteElement);
-    painter.restore(); // return painter to saved initial state
-}
-
-//---------------------------------------------------------
-//   paintStaff
-/// Paint a 5 line staff centered within a QRect and return the
-/// distance from the top of the QRect to the uppermost staff line.
-//---------------------------------------------------------
-
-qreal PaletteCellIconEngine::paintStaff(Painter& painter, const RectF& rect, qreal spatium)
-{
-    painter.save(); // so we can restore painter after we are done using it
-    Pen pen(configuration()->elementsColor());
-    pen.setWidthF(engraving::DefaultStyle::defaultStyle().value(Sid::staffLineWidth).toDouble() * spatium);
-    painter.setPen(pen);
-
-    constexpr int numStaffLines = 5;
-    const qreal staffHeight = spatium * (numStaffLines - 1);
-    const qreal topLineDist = rect.center().y() - (staffHeight / 2.0);
-
-    // lines bounded horizontally by edge of target (with small margin)
-    constexpr qreal margin = 3.0;
-    const qreal x1 = rect.left() + margin;
-    const qreal x2 = rect.right() - margin;
-
-    // draw staff lines with middle line centered vertically on target
-    qreal y = topLineDist;
-    for (int i = 0; i < numStaffLines; ++i) {
-        painter.drawLine(LineF(x1, y, x2, y));
-        y += spatium;
-    }
-
-    painter.restore(); // return painter to saved initial state
-    return topLineDist;
-}
-
-//---------------------------------------------------------
-//   paintBackground
-//---------------------------------------------------------
-
-void PaletteCellIconEngine::paintBackground(Painter& painter, const RectF& rect, bool selected, bool current)
-{
-    QColor c(configuration()->accentColor());
-    if (current || selected) {
-        c.setAlpha(selected ? 100 : 60);
-        painter.fillRect(rect, c);
-    }
-}
-
-//---------------------------------------------------------
-//   paintTag
-//---------------------------------------------------------
-
-void PaletteCellIconEngine::paintTag(Painter& painter, const RectF& rect, QString tag)
-{
-    if (tag.isEmpty()) {
-        return;
-    }
-
-    painter.save();   // so we can restore it after we are done using it
-    painter.setPen(configuration()->elementsColor());
-    Font f(painter.font());
-    f.setPointSizeF(12.0);
-    painter.setFont(f);
-
-    if (tag == "ShowMore") {
-        painter.drawText(rect, Qt::AlignCenter, "???");
-    } else {
-        painter.drawText(rect, Qt::AlignLeft | Qt::AlignTop, tag);
-    }
-
-    painter.restore();   // return to saved initial state (undo pen and font changes, etc.)
-}
-
-//---------------------------------------------------------
-//   PaletteCellIconEngine::paintCell
-//---------------------------------------------------------
-
-void PaletteCellIconEngine::paintCell(mu::draw::Painter& painter, const RectF& rect, bool selected, bool current) const
-{
-    const qreal _yOffset = 0.0;   // TODO
-
-    paintBackground(painter, rect, selected, current);
-
-    if (!_cell) {
-        return;
-    }
-
-    paintTag(painter, rect, _cell->tag);
-
-    Element* el = _cell->element.get();
-    if (!el) {
-        return;
-    }
-
-    if (el->isActionIcon()) {
-        paintIconElement(painter, rect, el);
-        return;     // never draw staff for icon elements
-    }
-
-    const bool drawStaff = _cell->drawStaff;
-    const qreal spatium = PALETTE_SPATIUM * _extraMag * _cell->mag;
-
-    PointF origin = rect.center();   // draw element at center of cell by default
-    painter.translate(0, _yOffset * spatium);   // offset both element and staff
-
-    if (drawStaff) {
-        const qreal topLinePos = paintStaff(painter, rect, spatium);     // draw dummy staff lines onto rect.
-        origin.setY(topLinePos);     // vertical position relative to staff instead of cell center.
-    }
-
-    painter.translate(origin);
-    painter.translate(_cell->xoffset * spatium, _cell->yoffset * spatium);   // additional offset for element only
-    painter.setPen(Pen(configuration()->elementsColor()));
-
-    paintScoreElement(painter, el, spatium, drawStaff);
-}
-
-//---------------------------------------------------------
-//   PaletteCellIconEngine::paint
-//---------------------------------------------------------
-
-void PaletteCellIconEngine::paint(QPainter* qp, const QRect& rect, QIcon::Mode mode, QIcon::State state)
-{
-    mu::draw::Painter p(qp, "palettecell");
-    p.save();   // so we can restore it later
-    p.setAntialiasing(true);
-    paintCell(p, RectF::fromQRectF(rect), mode == QIcon::Selected, state == QIcon::On);
-    p.restore();   // return painter to saved initial state (undo any changes to pen, coordinates, font, etc.)
-}
 } // namespace Ms
